#238 final revision for authentication demo & examples
