--- src/routes/demo/sign-in-illustration/+page.svelte
@@ -1,79 +1,0 @@
-<script lang="ts">
-	import {
-		El,
-		FormInput,
-		Button,
-		Card,
-		Divider,
-		Icon,
-		CardBody,
-		Label,
-		Input,
-		Checkbox,
-	} from 'yesvelte'
-
-	let icon = 'eye'
-	let textType = 'password'
-
-	const onClick = () => {
-		icon = icon === 'eye' ? 'eye-off' : 'eye'
-		textType = textType === 'password' ? 'text' : 'password'
-	}
-</script>
-
-<El class="page page-center" style="height:100vh;">
-	<El container="lg">
-		<El row m="0" alignItems="center" g="4">
-			<El colLg>
-				<El class="container-tight">
-					<El textAlign="center" mb="4">
-						<img src="/images/svelte.svg" height="36" alt="" class="me-2" />
-						<El tag="span" class="h1">YeSvelte</El>
-					</El>
-					<Card size="md">
-						<CardBody>
-							<El tag="h2" textAlign="center" mb="4">Login to your account</El>
-							<FormInput label="Email address" placeholder="your@email.com" />
-
-							<Label d="flex" justifyContent="between">
-								<El textAlign="start">Password</El>
-								<El textAlign="end">
-									<a href="./forgot-password.html">I forgot password</a>
-								</El>
-							</Label>
-							<Input type={textType} placeholder="Your password">
-								<Button tag="a" slot="end" on:click={onClick}><Icon name={icon} /></Button>
-							</Input>
-							<El mb="2" />
-							<Checkbox label="Remember me on this device" />
-
-							<Button color="primary" col="12" mt="4">Sign in</Button>
-
-							<Divider>OR</Divider>
-
-							<El row>
-								<El col="6">
-									<Button col="12">
-										<Icon name="brand-github" />
-										Login with Github
-									</Button>
-								</El>
-								<El col="6">
-									<Button col="12">
-										<Icon name="brand-twitter" color="twitter" />
-										Login with Twitter
-									</Button>
-								</El>
-							</El>
-						</CardBody>
-					</Card>
-					<El textAlign="center" mt="3"
-						>Don't have account yet? <a href="./sign-up.html" tabindex="-1">Sign up</a></El>
-				</El>
-			</El>
-			<El colLg="6" d="none" dLg="flex">
-				<El tag="img" src="/images/undraw_secure_login_pdn4.svg" height="300" mx="auto" />
-			</El>
-		</El>
-	</El>
-</El>
