+++ src/routes/docs/navigations.ts
@@ -402,32 +402,32 @@ export const navigations: Navigation[] = [
 		id: 'menu-examples',
 		children: [
 			{
-				route: example_path + '/sign-in',
+				route: example_path + '/authentication/sign-in',
 				title: 'Sign in',
 				icon: 'tabler:app-window',
 				description: 'A sample sign in page implemented with YeSvelte components.',
 			},
 			{
-				route: example_path + '/sign-in-illustration',
+				route: example_path + '/authentication/sign-in-illustration',
 				title: 'Sign in with illustration',
 				icon: 'tabler:app-window',
 				description:
 					'A sample sign in page with illustration implemented with YeSvelte components.',
 			},
 			{
-				route: example_path + '/sign-in-cover',
+				route: example_path + '/authentication/sign-in-cover',
 				title: 'Sign in with cover',
 				icon: 'tabler:app-window',
 				description: 'A sample sign in page with cover image implemented with YeSvelte components.',
 			},
 			{
-				route: example_path + '/sign-up',
+				route: example_path + '/authentication/sign-up',
 				title: 'Sign up',
 				icon: 'tabler:app-window',
 				description: 'A sample sign up page implemented with YeSvelte components.',
 			},
 			{
-				route: example_path + '/forgot-password',
+				route: example_path + '/authentication/forgot-password',
 				title: 'Forgot password',
 				icon: 'tabler:app-window',
 				description: 'A sample forgot password page implemented with YeSvelte components.',
